Alpaca/requirements.txt
pytest==3.2.1
numpy
gensim
h5py
requests
sklearn
neo4j-driver
flask
flask-restful
flask-api
flask-cors
